# Conflicts:
#	TeamCode/src/main/java/org/firstinspires/ftc/teamcode/commands/Teleop/GrabberCommand.java

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/GrabberSubsystem.java
@@ -36,7 +36,7 @@ public class GrabberSubsystem {
             grabberClosed = true;
         }
     }
-    public void setGrabberPosition (int grabberPosition) {
+    public void setGrabberPosition (double grabberPosition) {
         hardware.grabber.setPosition(grabberPosition);
     }
